# src/swagger/swagger.yml
openapi: 3.0.0
info:
  title: MineBot API with TypeScript
  description: This is the API documentation for the Express application built with TypeScript.
  version: 1.0.0
  contact:
    name: Your Name
    email: your.email@example.com
    url: https://yourwebsite.com
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Development server
  - url: 'http://35.77.96.67:3000/api/v1'
    description: Live server
openapi: 3.0.0
info:
  title: Express API with TypeScript
  description: This is the API documentation for the Express application built with TypeScript.
  version: 1.0.0
  contact:
    name: Your Name
    email: your.email@example.com
    url: https://yourwebsite.com
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Development server
  - url: 'http://35.77.96.67:3000/api/v1'
    description: Live server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /reset/forgot-password:
    post:
      summary: Send a password reset email to the provided email address
      operationId: forgotPassword
      tags:
        - Forgot Password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: The email address of the user requesting a password reset
      responses:
        '200':
          description: Reset email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reset email sent"
        '400':
          description: Bad request, missing required fields or invalid data
        '500':
          description: Internal server error
  /reset/reset-password:
    get:
      summary: Reset the user's password using the provided token and new password
      operationId: resetPassword
      tags:
        - Reset Password
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
            description: The token received in the password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  description: The new password for the user
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successful"
        '400':
          description: Bad request, missing required fields or invalid data
        '401':
          description: Unauthorized, invalid or expired token
        '500':
          description: Internal server error
...